LayoutTests/imported/w3c/web-platform-tests/referrer-policy/gen/srcdoc-inherit.http-rp/origin/xhr.http-expected.txt

Blocked access to external URL http://www1.localhost:8800/common/security-features/subresource/xhr.py?redirection=keep-origin&action=purge&key=3a73d1d4-b0a6-4a19-a33e-d0b0bbb7ec67&path=%2Fmixed-content
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.localhost:8800/common/security-features/subresource/xhr.py?redirection=keep-origin&action=purge&key=3a73d1d4-b0a6-4a19-a33e-d0b0bbb7ec67&path=%2Fmixed-content due to access control checks.
Blocked access to external URL http://www1.localhost:8800/common/security-features/subresource/xhr.py?redirection=no-redirect&action=purge&key=770db698-437e-46a2-8500-eb8169e5a8ea&path=%2Fmixed-content
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.localhost:8800/common/security-features/subresource/xhr.py?redirection=no-redirect&action=purge&key=770db698-437e-46a2-8500-eb8169e5a8ea&path=%2Fmixed-content due to access control checks.
Blocked access to external URL http://www1.localhost:8800/common/security-features/subresource/xhr.py?redirection=swap-origin&action=purge&key=179e24be-d754-4aa0-afde-98595e47ff37&path=%2Fmixed-content
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.localhost:8800/common/security-features/subresource/xhr.py?redirection=swap-origin&action=purge&key=179e24be-d754-4aa0-afde-98595e47ff37&path=%2Fmixed-content due to access control checks.
Blocked access to external URL https://www1.localhost:9443/common/security-features/subresource/xhr.py?redirection=keep-origin&action=purge&key=fc33d011-53e5-482e-b33b-862b8bfcf6b1&path=%2Fmixed-content
CONSOLE MESSAGE: XMLHttpRequest cannot load https://www1.localhost:9443/common/security-features/subresource/xhr.py?redirection=keep-origin&action=purge&key=fc33d011-53e5-482e-b33b-862b8bfcf6b1&path=%2Fmixed-content due to access control checks.
Blocked access to external URL https://www1.localhost:9443/common/security-features/subresource/xhr.py?redirection=no-redirect&action=purge&key=524735e6-bd82-490e-b3c3-d206e91f9655&path=%2Fmixed-content
CONSOLE MESSAGE: XMLHttpRequest cannot load https://www1.localhost:9443/common/security-features/subresource/xhr.py?redirection=no-redirect&action=purge&key=524735e6-bd82-490e-b3c3-d206e91f9655&path=%2Fmixed-content due to access control checks.
Blocked access to external URL https://www1.localhost:9443/common/security-features/subresource/xhr.py?redirection=swap-origin&action=purge&key=967deb96-32eb-4036-bd04-fb061a10f6be&path=%2Fmixed-content
CONSOLE MESSAGE: XMLHttpRequest cannot load https://www1.localhost:9443/common/security-features/subresource/xhr.py?redirection=swap-origin&action=purge&key=967deb96-32eb-4036-bd04-fb061a10f6be&path=%2Fmixed-content due to access control checks.
Blocked access to external URL http://www1.localhost:8800/common/security-features/subresource/xhr.py?action=purge&key=33b2d5cb-0861-4493-8fc9-484a4892941c&path=%2Fmixed-content
Blocked access to external URL https://www1.localhost:9443/common/security-features/subresource/xhr.py?action=purge&key=8751da19-e7a6-49b7-8e3c-f25b898012dc&path=%2Fmixed-content

FAIL Referrer Policy: Expects origin for xhr to cross-http origin and keep-origin redirection from http context. promise_test: Unhandled rejection with value: "Network Error"
FAIL Referrer Policy: Expects origin for xhr to cross-http origin and no-redirect redirection from http context. promise_test: Unhandled rejection with value: "Network Error"
FAIL Referrer Policy: Expects origin for xhr to cross-http origin and swap-origin redirection from http context. promise_test: Unhandled rejection with value: "Network Error"
FAIL Referrer Policy: Expects origin for xhr to cross-https origin and keep-origin redirection from http context. promise_test: Unhandled rejection with value: "Network Error"
FAIL Referrer Policy: Expects origin for xhr to cross-https origin and no-redirect redirection from http context. promise_test: Unhandled rejection with value: "Network Error"
FAIL Referrer Policy: Expects origin for xhr to cross-https origin and swap-origin redirection from http context. promise_test: Unhandled rejection with value: "Network Error"
PASS Referrer Policy: Expects origin for xhr to same-http origin and keep-origin redirection from http context.
PASS Referrer Policy: Expects origin for xhr to same-http origin and no-redirect redirection from http context.
FAIL Referrer Policy: Expects origin for xhr to same-http origin and swap-origin redirection from http context. promise_test: Unhandled rejection with value: "Network Error"
PASS Referrer Policy: Expects origin for xhr to same-https origin and keep-origin redirection from http context.
PASS Referrer Policy: Expects origin for xhr to same-https origin and no-redirect redirection from http context.
FAIL Referrer Policy: Expects origin for xhr to same-https origin and swap-origin redirection from http context. promise_test: Unhandled rejection with value: "Network Error"
